requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
factor_analyzer>=0.5.0
openai>=1.0.0
google-generativeai>=0.8.0
flask>=3.0.0
tabulate>=0.9.0
